Separate thrift package loading related functionality to a new class: ThriftLoader
