+++ thriftcli/thrift_argument_converter.py
@@ -20,9 +20,8 @@ from .thrift_parser import ThriftParser
 class ThriftArgumentConverter(object):
     """ Converts a json request body into the corresponding Python object generated by thrift. """
 
-    def __init__(self, thrift_path, thrift_dir_paths=None):
-        thrift_parser = ThriftParser(thrift_path, thrift_dir_paths)
-        self._parse_result = thrift_parser.parse()
+    def __init__(self, result):
+        self._parse_result = result
 
     def convert_args(self, service_reference, method_name, data):
         """ Converts json request body into keyword arguments for a service's method.
