+++ thriftcli/thrift_executor.py
@@ -27,29 +27,19 @@ from .thrift_cli_error import ThriftCLIError
 class ThriftExecutor(object):
     """ This class handles connecting to and communicating with the Thrift server. """
 
-    def __init__(self, thrift_path, server_address, service_reference, basename_to_namespaces, thrift_dir_paths=None,
-                 client_id=None):
+    def __init__(self, server_address, service_name, thrift_loader, client_id=None):
         """ Opens a connection with the server and generates then imports the thrift-defined python code.
 
-        :param thrift_path: the path to the Thrift file defining the service being requested
         :param server_address: the address to the server implementing the service
-        :param service_reference: the namespaced service name in the format <file-name>.<service-name>
-        :param thrift_dir_paths: a list of paths to directories containing Thrift file dependencies
         :param client_id: Finagle client id for identifying requests
 
         """
-        self._thrift_path = thrift_path
         self._server_address = server_address
-
-        self._thrift_dir_paths = set(thrift_dir_paths) if thrift_dir_paths is not None else set([])
-        # Handle case where thrift file is in the current directory
-        thrift_file_dir = os.path.dirname(thrift_path) or '.'
-        self._thrift_dir_paths.add(thrift_file_dir)
+        self._service_name = service_name
+        self._thrift_loader = thrift_loader
 
         self._client_id = client_id
-        self._service_reference = service_reference
         self._open_connection(server_address)
-        self._generate_and_import_packages(basename_to_namespaces)
 
     def run(self, method_name, request_args):
         """ Executes a method on the connected server and returns its result.
@@ -61,7 +51,7 @@ class ThriftExecutor(object):
         :return: the result of the method call
 
         """
-        method = self._get_method(method_name)
+        method = self._thrift_loader.get_client_method(service_name, method_name)
         return method(**request_args)
 
     def cleanup(self, remove_generated_src=False):
@@ -70,56 +60,9 @@ class ThriftExecutor(object):
         :param remove_generated_src: whether or not to delete the generated source
 
         """
-        if remove_generated_src:
-            self._remove_dir('gen-py')
+        self._thrift_loader.cleanup(remove_generated_src)
         if self._transport:
             self._transport.close()
-
-    @staticmethod
-    def _remove_dir(path):
-        """ Recursively removes a directory and ignores if it didn't exist.
-
-        :param path: the directory to remove
-
-        """
-        try:
-            shutil.rmtree(path)
-        except OSError:
-            pass
-
-    def _generate_and_import_packages(self, basename_to_namespaces):
-        """ Generates and imports the python modules defined by the thrift code.
-
-        This method does the following:
-        1. Runs a shell process to generate the python code from the Thrift file
-        2. Adds the generated source to the python process' path
-        3. Imports the generated source package into this python process
-
-        """
-        thrift_dir_options = ''.join([' -I %s' % thrift_dir_path for thrift_dir_path in self._thrift_dir_paths])
-        command = 'thrift -r%s --gen py %s' % (thrift_dir_options, self._thrift_path)
-        if subprocess.call(command, shell=True) != 0:
-            raise ThriftCLIError('Thrift generation command failed: \'%s\'' % command)
-        sys.path.append('gen-py')
-        for basename, package in basename_to_namespaces.items():
-            self._import_package(basename, package)
-
-    def _get_method(self, method_name):
-        """ Returns the python method generated for the given endpoint.
-
-        :param method_name: the name of the method to retrieve
-        :returns: the python method that can be called to execute the Thrift RPC
-        :rtype: method
-
-        """
-        class_name = 'Client'
-        client_constructor = getattr(sys.modules[self._service_reference], class_name)
-        client = client_constructor(self._protocol)
-        try:
-            method = getattr(client, method_name)
-        except AttributeError:
-            raise ThriftCLIError('\'%s\' service has no method \'%s\'' % (self._service_reference, method_name))
-        return method
 
     def _open_connection(self, address):
         """ Opens a connection with a server address.
@@ -147,16 +90,3 @@ class ThriftExecutor(object):
         url_obj = urlparse.urlparse(address)
         return url_obj.hostname, url_obj.port
 
-    @staticmethod
-    def _import_package(basename, package_name):
-        """ Imports a package generated by thrift code 
-
-        :param package_name: the name of the package to import, which must be located somewhere on sys.path
-
-        """
-        package = importlib.import_module(package_name)
-        for module in package.__all__:
-            sub_module_name = '.'.join([basename, module])
-            sub_package_name = '.'.join([package_name, module])
-            sub_package = importlib.import_module(sub_package_name)
-            sys.modules[sub_module_name] = sub_package
